+++ src/game/game_init.h
@@ -43,7 +43,7 @@ extern struct GfxPool *gGfxPool;
 extern u8 gControllerBits;
 extern u8 gIsConsole;
 #ifdef WIDE
-extern u8 gWidescreen;
+extern s16 gWidescreen;
 #endif
 extern u8 gBorderHeight;
 #ifdef CUSTOM_DEBUG
